Issue #13234: page.includeJs() callbacks should fire only once.
Thanks to the awkward semantics of JS 'this', callbacks registered for
page.includeJs() completion were not getting deregistered.  This was
invisible under most conditions, but: if you loaded a page, included JS
into it, loaded another page using the same webpage object, and then
included the *same* JS file into the second page, the callback you
provided for the *first* call to includeJs would get called a second
time.